+++ src/components/select/ValueContainer.tsx
@@ -1,0 +1,39 @@
+import { ReactSelect } from '@getdokan/dokan-ui';
+
+// Render a single-line summary for multi-select instead of chips
+const ValueContainer = ( valueProps: any ) => {
+    const { components } = ReactSelect;
+    const { selectProps, hasValue } = valueProps as any;
+    const isMulti = Boolean( selectProps?.isMulti );
+    if ( ! isMulti ) {
+        return <components.ValueContainer { ...valueProps } />;
+    }
+    const values: any[] = ( selectProps?.value as any[] ) || [];
+    const labels = values.map( ( v ) => v?.label ?? v ).filter( Boolean );
+    const titlePrefix = selectProps?.selectedTitle;
+    const prefix =
+        typeof titlePrefix === 'function' && labels.length
+            ? titlePrefix( values[ 0 ] )
+            : titlePrefix;
+    const summary = labels.join( ', ' );
+    // Keep the input so user can type; hide chips via MultiValue override
+    return (
+        <components.ValueContainer { ...valueProps }>
+            { hasValue ? (
+                <div className="break-all text-[14px] leading-[22px] text-gray-700">
+                    { prefix ? (
+                        <span className="font-normal">
+                            { `${ String( prefix ) }: ` }
+                        </span>
+                    ) : null }
+                    <span className="align-middle" title={ summary }>
+                        { summary }
+                    </span>
+                </div>
+            ) : null }
+            { valueProps.children }
+        </components.ValueContainer>
+    );
+};
+
+export default ValueContainer;
